Couch2pg added better logging

+++ bilanci_project/bilanci/management/commands/couch2pg.py
@@ -374,8 +374,9 @@ class Command(BaseCommand):
         else:
             self.logger.info("Deleting values for selected cities, years")
 
-        if not self.dryrun:
+        if not self.dryrun and values_to_delete.count()>0:
             values_to_delete.delete()
+        self.logger.info("Done deleting")
 
         # creates somma_funzioni_slug_baseset
         for slug in self.considered_somma_funzioni:
